+++ Proyect3TWN.WebSite/Pages/VacationsAbsences/Delete.cshtml.cs
@@ -9,12 +9,18 @@ public class Delete : PageModel
 {
     [BindProperty] public VacationsAbsencesDto VacationsAbsencesDto { get; set; }
 
+    
+    public Dictionary<int, string> EmployeesName { get; set; }
     public List<string> Errors { get; set; } = new List<string>();
     private readonly IVacationsAbsencesService _service;
+    private readonly IEmployeeService _employeeService;
 
-    public Delete(IVacationsAbsencesService service)
+    public Delete(IVacationsAbsencesService service, IEmployeeService employeeService)
     {
         _service = service;
+        _employeeService = employeeService;
+        EmployeesName = new Dictionary<int, string>();
+
     }
     public async Task<IActionResult> OnGet(int id)
     {
@@ -27,11 +33,21 @@ public class Delete : PageModel
         }
         VacationsAbsencesDto = response.Data;
 
+        var employeeResponse = await _employeeService.GetAllAsync();
+        if (employeeResponse?.Data != null)
+        {
+            EmployeesName = employeeResponse.Data.ToDictionary(d => d.id, d => d.Name);
+        }
+        
         return Page();
     }
 
     public async Task<IActionResult> OnPostAsync()
     {
+        if (VacationsAbsencesDto == null)
+        {
+            return NotFound();
+        }
         var response = await _service.DeleteAsync(VacationsAbsencesDto.id);
         return RedirectToPage("./List");
     }
